Use Newer TypedData_* Methods For Object Access
Previous we were using the old-style Data_* methods. These methods do not
perform any type validations and this makes it easy to SIGSEGV a program
using this extension gem simply by passing arguments to functions in the
wrong order.

The code base has now been completely migrated to the TypedData_* methods. This
required some reorganization as well as the definition of `rb_data_type_t`
structures for all objects provided by the library. In addition, allocation and
deallocation of memory now uses the built-in methods and objects are
appropriately marked as ready for GC at the end of their lifecycle. This should
further alleviate any bugs / memory-leaks in the previous release.