Copy Contexts Onto Objects
Currently, we copy around the `Context` struct rather than just cloning the
`secp256k1_context` itself. This probably causes segmentation faults when
object lifetimes are not closely aligned with those of the `Context` object
that created them.

This change modifies this so that contexts are cloned onto objects that require
them for instance methods. This makesso context lifetimes safer (e.g. objects
using a context cannot outlive it).