Upgrade bitcoin-core/secp256k1 to latest HEAD
`rbspec256k1` is not running on m1 macs. We're seeing this error: https://github.com/etscrivner/rbsecp256k1/issues/54

I dug into the depedencies and noticed that the bitcoin-core/secp256k1 depedency has not been update in 3 years. Using the latest `HEAD` fixed the issue on m1 macs and all tests are passing.