[Backwards Incompatible] Refactor KeyPair to use secp256k1_keypair
Refactor the `KeyPair` class to use the `secp256k1_keypair` type and its
associated methods. This new type is required by schnorrsig and is largely a
drop-in replacement for the functionality previously provided by `KeyPair`.

Remove `KeyPair#initialize` since it no longer correctly mirrors the internal
structure of the class. This breaks backwards compatibility.